feat: add new 'runtime' option to target electron releases
As of Electron 4.0.0 we probably won't be shipping node.lib or iojs.lib
we will instead be shipping electron.lib.

In order for node-gyp to target our *.lib file correctly we need to add
an option to target electron.lib instead of node.lib.  We can't blanket
do this as we still ship node.lib for <= 3.X.

As such I've added this runtime flag and a semver check to target
electron.lib for only versions 4.0.0-a.0 and higher.  The `.a-0` is to
handle our prerelease tags (beta, nightly) and still correctly return
`true` for `semver.gte`.

We're a while away from shipping 4.x but I'd like to get this reviewed
ASAP so it's already out there and ready to use when we get to 4.x ship
date.